feat(form-elements): Add image import functionality and update tool panel builder
A new ImageImport component is created in form-elements. This component reads and previews the file selected by the user. During file loading, errors are handled and displayed in an ErrorMessage component. Styling is applied to properly align the elements.

In the ErrorMessage component, minor adjustments to the timing and display properties were made to handle display of errors. Additional styling was also included to improve the display position of error messages.

In Icon component, an inline property is added to control the styling of the icons.

A new Spinner component for displaying loading states was created, utilizing the Dot Linear Shift spinner from loadership.com.

ToolPanelBuilder is updated to utilize the new ImageImport component.

CSS updates involved importing new buttons.css for button styling and adjusting some file imports in app.css. New specific styles for buttons added in buttons.css.

The bootstrapIcons resource file was updated with an additional upload icon, which was utilized in the new ImageImport component.