refactor(form-elements): Update form element configuration naming and add image importing
The naming of the configuration types for the Checkbox, Slider, and Switch components have been updated. 'Configs' has been replaced with 'Config'. This change is reflected in all components and utility functions where these configuration types are used, leading to a more consistent codebase.

A configuration for a new form element, ImageImport, is also added. This configuration type is incorporated into the union 'ConfigType'. The ImageImport component is also added to potential form elements handled in the ToolPanelBuilder, thereby allowing it to utilize this new component.